+++ packages/sources/src/components/modal/modal.models.ts
@@ -1,10 +1,14 @@
-import { Button } from "../button/button.models";
-
-export interface Modal<TemplateFnReturnType> {
-  id?: string;
-  heading?: string;
-  role: "alert" | "dialog";
-  body: TemplateFnReturnType | string;
-  buttons?: Button[];
-  confirm?: boolean;
+export interface Modal {
+  modalTitle?: string;
+  body: string;
+  footer?: string;
+  role: ModalRole;
+  dsoClose?: (value: CustomEvent) => void;
 }
+
+export interface ModalMethods {
+  openModal(): Promise<void>;
+  closeModal(): Promise<void>;
+}
+
+export type ModalRole = "alert" | "dialog" | "alertdialog";